Add shared cross-lingual translator module
