+++ src/reasoning_history.py
@@ -12,7 +12,7 @@ if __name__ not in sys.modules:  # pragma: no cover - for manual loaders
     sys.modules[__name__] = types.ModuleType(__name__)
 
 if TYPE_CHECKING:  # pragma: no cover - only for type hints
-    from .data_ingest import CrossLingualTranslator
+    from .cross_lingual_translator import CrossLingualTranslator
     from .graph_of_thought import GraphOfThought
     from .graph_pruning_manager import GraphPruningManager
